Make the shared selection colors sane.
The monokai theme does not work well if the text-
background colors are very bright. It makes the
selected text very hard to read for the other user
in the same editor.

This uses darker muted versions of some monokai
colors, and assigns one at random to each user.

This could cause 2 users to have the same cursor
color, but unless there are 3+ simultaneous
editors, nobody will notice.